+++ component/frontend/ViewTemplates/Categories/customrepo.blade.php
@@ -17,6 +17,8 @@
 $customHTML = @file_get_contents($this->customHtmlFile);
 
 ?>
+@include('site:com_ars/Categories/chromeisbroken')
+
 @unless(empty($customHTML))
     {{ Joomla\CMS\HTML\HTMLHelper::_('content.prepare', $customHTML) }}
 @else